+++ app/Http/Controllers/Auth/TwoFactorController.php
@@ -1,9 +1,10 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Auth;
 
 use App\Models\User;
 use Illuminate\Http\Request;
+use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Auth;
 use App\Http\Requests\VerifyTwoFactorRequest;
 use Illuminate\Foundation\Auth\AuthenticatesUsers;
