+++ routes/web.php
@@ -14,7 +14,6 @@ use App\Http\Controllers\CustomerController;
 use App\Http\Controllers\LocationController;
 use App\Http\Controllers\KpiReportController;
 use App\Http\Controllers\SocialiteController;
-use App\Http\Controllers\TwoFactorController;
 use App\Http\Controllers\DepartmentController;
 use App\Http\Controllers\HrCalendarController;
 use App\Http\Controllers\DesignationController;
@@ -24,6 +23,7 @@ use App\Http\Controllers\OfficeShiftController;
 use App\Http\Controllers\ServiceFormController;
 use App\Http\Controllers\AnnouncementController;
 use App\Http\Controllers\KpiObjectiveController;
+use App\Http\Controllers\Auth\TwoFactorController;
 use App\Http\Controllers\ChangePasswordController;
 use App\Http\Controllers\MachineRequestController;
 use App\Http\Controllers\AcknowledgementFormController;
